Move BlockRetriever to blocks folder
