+++ casper/src/test/scala/coop/rchain/casper/sync/BlockRetrieverRequesAllSpec.scala
@@ -2,9 +2,8 @@ package coop.rchain.casper.sync
 
 import cats.effect.concurrent.Ref
 import com.google.protobuf.ByteString
-import coop.rchain.casper.engine
-import coop.rchain.casper.engine.BlockRetriever
-import coop.rchain.casper.engine.BlockRetriever.RequestState
+import coop.rchain.casper.blocks.BlockRetriever
+import coop.rchain.casper.blocks.BlockRetriever.{RequestState, RequestedBlocks}
 import coop.rchain.casper.protocol._
 import coop.rchain.catscontrib.ski._
 import coop.rchain.comm.CommError._
@@ -34,7 +33,7 @@ class BlockRetrieverRequestAllSpec extends AnyFunSpec with BeforeAndAfterEach wi
 
   implicit val log     = new LogStub[Task]
   implicit val metrics = new Metrics.MetricsNOP[Task]
-  implicit val currentRequests: engine.BlockRetriever.RequestedBlocks[Task] =
+  implicit val currentRequests: RequestedBlocks[Task] =
     Ref.unsafe[Task, Map[BlockHash, RequestState]](Map.empty[BlockHash, RequestState])
   implicit val connectionsCell: ConnectionsCell[Task] =
     Ref.unsafe[Task, Connections](List(local))
