+++ casper/src/test/scala/coop/rchain/casper/sync/BlockRetrieverSpec.scala
@@ -2,13 +2,12 @@ package coop.rchain.casper.sync
 
 import cats.effect.concurrent.Ref
 import com.google.protobuf.ByteString
-import coop.rchain.casper.engine
-import coop.rchain.casper.engine.BlockRetriever
-import coop.rchain.casper.engine.BlockRetriever.RequestState
+import coop.rchain.casper.blocks.BlockRetriever
+import coop.rchain.casper.blocks.BlockRetriever.{RequestState, RequestedBlocks}
 import coop.rchain.casper.protocol._
-import coop.rchain.comm.{Endpoint, NodeIdentifier, PeerNode}
 import coop.rchain.comm.rp.Connect.{Connections, ConnectionsCell}
 import coop.rchain.comm.rp.ProtocolHelper._
+import coop.rchain.comm.{Endpoint, NodeIdentifier, PeerNode}
 import coop.rchain.metrics.Metrics
 import coop.rchain.models.BlockHash.BlockHash
 import coop.rchain.p2p.EffectsTestInstances.{createRPConfAsk, LogStub, TransportLayerStub}
@@ -34,7 +33,7 @@ class BlockRetrieverSpec extends AnyFunSpec with BeforeAndAfterEach with Matcher
 
   implicit val log     = new LogStub[Task]
   implicit val metrics = new Metrics.MetricsNOP[Task]
-  implicit val currentRequests: engine.BlockRetriever.RequestedBlocks[Task] =
+  implicit val currentRequests: RequestedBlocks[Task] =
     Ref.unsafe[Task, Map[BlockHash, RequestState]](Map.empty[BlockHash, RequestState])
   implicit val connectionsCell: ConnectionsCell[Task] =
     Ref.unsafe[Task, Connections](List(local))
